feat: add modal box for event
