+++ src/components/common/modal/index.ts
@@ -1,0 +1,1 @@
+export { ModalBox } from './Modal';
