+++ src/components/common/modal/modal.module.css
@@ -1,0 +1,16 @@
+.modalContainer {
+    position: absolute;
+    width: 50%;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    display: flex;
+    padding: 24px;
+    flex-direction: column;
+    align-items: flex-start;
+    border-radius: 8px;
+    border: 1px solid #c896ef;
+    background: #201b27;
+    color: #fff;
+    outline: none;
+}
