+++ src/components/pages/RequestPage/RequestPage.tsx
@@ -1,8 +1,13 @@
 import { useState } from 'react';
-import { RequestModal } from './components/Modal';
+import SimpleCloseIcon from '../../../assets/SupportServicePanel/SimpleCloseIcon.svg';
+import FileIcon from '../../../assets/SupportServicePanel/FileIcon.svg';
 import SearchInput from './components/SearchInput';
 import RequestTable from './components/Table';
 import styles from './request.module.css';
+import { ModalBox } from 'components/common/modal';
+import { Table } from './components/types';
+import { tableData } from './components/table-data';
+import { TextareaAutosize } from '@mui/material';
 
 export function RequestPage() {
   const [open, setOpen] = useState(false);
@@ -12,6 +17,10 @@ export function RequestPage() {
   const handleChangeOpen = () => {
     setOpen(!open);
   };
+
+  const filteredDataById: Table | undefined = tableData.find(
+    (item) => item.id === id,
+  );
 
   return (
     <div className={styles.container}>
@@ -26,7 +35,57 @@ export function RequestPage() {
         setId={setId}
         text={text}
       />
-      <RequestModal open={open} handleChangeOpen={handleChangeOpen} id={id} />
+      <ModalBox open={open} handleChangeOpen={handleChangeOpen}>
+        <div className={styles.modalHeader}>
+          <div className={styles.modalHeaderLeft}>
+            <p>{filteredDataById?.subject}</p>
+            <div
+              className={styles.status}
+              style={{
+                backgroundColor: `${filteredDataById?.backGroundColor}`,
+              }}
+            >
+              {filteredDataById?.status}
+            </div>
+          </div>
+          <img
+            src={SimpleCloseIcon}
+            alt="SimpleCloseIcon"
+            style={{
+              cursor: 'pointer',
+            }}
+            onClick={handleChangeOpen}
+          />
+        </div>
+        <div className={styles.idSection}>
+          <div>
+            <p>Ticket ID:</p>
+            <p>{filteredDataById?.id}</p>
+          </div>
+          <div>
+            <p>Category:</p>
+            <p>Software</p>
+          </div>
+        </div>
+        <div className={styles.description}>
+          <p>Description:</p>
+          <TextareaAutosize
+            value="Software flgkd;lgkh;lkg fmkdlls;;as.d,fmfngmd,s; dlgmmfdfmgfmmdfsfkjgkrjgjfgksfjgf;jkjgiri frjrigjfndmkkeigi dkfgfkdmfkfefnwklrn;wkrjnmSoftwarelgkd;lgkh;lkgfmkdlls;;as.d,fmfngmd,s;dlgmmfdfmgfmmdfsfkjgkrjgjfgksfjgf;jkjgiri frjrigjfndmkkeigi dkfgfk dmfkf efnwklrn;wkrjn mSoftware flgkd;lgkh;lkg fmkdlls;;as.d,fmfngmd,s; ghfjfjkfkfkfkfkfkfkdlgmmfdfmgfmmdfsfkjgkrjgjfgksfjgf;jkjgiri  frjrigjfndmkkeigi dkfgfk dmfkf efnwklrn;wkrjn m"
+            disabled
+          />
+        </div>
+        <div className={styles.files}>
+          <p>Files included:</p>
+          <div>
+            <img src={FileIcon} alt="" />
+            <span>Issue.jpg</span>
+          </div>
+          <div>
+            <img src={FileIcon} alt="" />
+            <span>Issue.pdf</span>
+          </div>
+        </div>
+      </ModalBox>
     </div>
   );
 }
